one more fix

+++ docker-images/scripts/unzip.sh
@@ -4,4 +4,8 @@ then
   echo "$0: you have to specify unzip password via ZIPPASS=..."
   exit 1
 fi
-unzip -P "$ZIPPASS" data.zip > /dev/null && rm -f data.zip
+
+if [ -f "data.zip" ]
+then
+  unzip -P "$ZIPPASS" data.zip > /dev/null && rm -f data.zip
+fi
